+++ Spot.IM-Development/SpotImCore/Core/Models/SPCommentViewModel.swift
@@ -27,8 +27,8 @@ internal struct CommentViewModel {
     var commentLabel: CommentLabel?
     var commentGifUrl: String?
     var commentImage: CommentImage?
-    var commentMediaHeight: Float?
-    var commentMediaWidth: Float?
+    private var commentMediaOriginalHeight: Int?
+    private var commentMediaOriginalWidth: Int?
     
     var replyingToDisplayName: String?
     var replyingToCommentId: String?
@@ -83,12 +83,14 @@ internal struct CommentViewModel {
         
         if let gif = comment.gif {
             commentGifUrl = gif.originalUrl
-            (self.commentMediaHeight, self.commentMediaWidth) = self.calculateMediaSize(mediaHeight: gif.previewHeight, mediaWidth: gif.previewWidth)
+            self.commentMediaOriginalHeight = gif.previewHeight
+            self.commentMediaOriginalWidth = gif.previewWidth
         }
         
         if let image = comment.image, let commentImageURL = commentImageURL {
             commentImage = CommentImage(id: image.imageId, height: image.originalHeight, width: image.originalWidth, imageUrl: commentImageURL)
-            (self.commentMediaHeight, self.commentMediaWidth) = self.calculateMediaSize(mediaHeight: image.originalHeight, mediaWidth: image.originalWidth)
+            self.commentMediaOriginalHeight = image.originalHeight
+            self.commentMediaOriginalWidth = image.originalWidth
         }
         
         
@@ -152,14 +154,17 @@ internal struct CommentViewModel {
 
     func textWidth() -> CGFloat {
         let leadingOffset: CGFloat = depthOffset()
-        let textWidth = UIScreen.main.bounds.width - leadingOffset - Theme.trailingOffset
+        let textWidth = SPUIWindow.frame.width - leadingOffset - Theme.trailingOffset
         
         return textWidth
     }
     
-    func calculateMediaSize(mediaHeight: Int, mediaWidth: Int) -> (Float, Float) {
+    func getMediaSize() -> CGSize {
+        guard let mediaHeight = commentMediaOriginalHeight,
+              let mediaWidth = commentMediaOriginalWidth
+        else { return CGSize(width: 0, height: 0) }
         let leadingOffset: CGFloat = depthOffset()
-        let maxWidth = UIScreen.main.bounds.width - leadingOffset - Theme.trailingOffset
+        let maxWidth = SPUIWindow.frame.width - leadingOffset - Theme.trailingOffset
         
         // calculate media width according to height ratio
         var height = Theme.commentMediaMaxHeight
@@ -172,7 +177,7 @@ internal struct CommentViewModel {
             height = (ratio * Float(mediaHeight))
         }
         
-        return (height, width)
+        return CGSize(width: CGFloat(width), height: CGFloat(height))
     }
     
     func height(with lineLimit: Int, isLastInSection: Bool = false) -> CGFloat {
@@ -187,7 +192,7 @@ internal struct CommentViewModel {
             0.0 : clippedMessage.height(withConstrainedWidth: width)
         
         // media extra height includes - media acual heigh + media extra padding
-        let mediaHeight: CGFloat = commentMediaHeight == nil ? 0.0 : CGFloat(commentMediaHeight! + Float(SPCommonConstants.commentMediaTopPadding - SPCommonConstants.emptyCommentMediaTopPadding))
+        let mediaHeight = CGFloat(Float(getMediaSize().height) + Float(SPCommonConstants.commentMediaTopPadding - SPCommonConstants.emptyCommentMediaTopPadding))
         
         let moreRepliesHeight = repliesButtonState == .hidden ?
             0.0 : Theme.moreRepliesViewHeight + Theme.moreRepliesTopOffset