+++ Spot.IM-Development/SpotImCore/SpotIMSDK/UI/ViewControllers/MainConversation/SPMainConversationModel.swift
@@ -22,12 +22,12 @@ typealias BooleanCompletion = (Bool, Error?) -> Void
 
 let navigationAvatarSize: CGSize = CGSize(width: 25.0, height: 25.0)
 
-protocol MainConversationModelDelegate: class {
+protocol MainConversationModelDelegate: AnyObject {
     func totalTypingCountDidUpdate(count: Int)
     func stopTypingTrack()
 }
 
-protocol CommentsCounterDelegate: class {
+protocol CommentsCounterDelegate: AnyObject {
     func commentsCountDidUpdate(count: Int)
 }
 
@@ -187,20 +187,20 @@ final class SPMainConversationModel {
     func commentAvailableActions(_ commentId: String, sender: UIButton) -> [UIAlertAction] {
         let viewModel = dataSource.commentViewModel(commentId)
         let availability = commentActionsAvailability(viewModel: viewModel)
-        
+        let replyingToID = viewModel?.rootCommentId
         var actions: [UIAlertAction] = []
 
         let shareAction = UIAlertAction(
             title: LocalizationManager.localizedString(key: "Share"),
             style: .default) { [weak self] _ in
-                self?.commentsActionDelegate?.prepareFlowForAction(.share(commentId: commentId), sender: sender)
+            self?.commentsActionDelegate?.prepareFlowForAction(.share(commentId: commentId, replyingToID: replyingToID), sender: sender)
         }
         actions.append(shareAction)
         if availability.isReportable {
             let reportAction = UIAlertAction(
                 title: LocalizationManager.localizedString(key: "Report"),
                 style: .default) { [weak self] _ in
-                    self?.commentsActionDelegate?.prepareFlowForAction(.report(commentId: commentId), sender: sender)
+                self?.commentsActionDelegate?.prepareFlowForAction(.report(commentId: commentId, replyingToID: replyingToID), sender: sender)
             }
             actions.append(reportAction)
         }
@@ -209,7 +209,7 @@ final class SPMainConversationModel {
             let deleteAction = UIAlertAction(
                 title: LocalizationManager.localizedString(key: "Delete"),
                 style: .default) { [weak self] _ in
-                    self?.commentsActionDelegate?.prepareFlowForAction(.delete(commentId: commentId), sender: sender)
+                self?.commentsActionDelegate?.prepareFlowForAction(.delete(commentId: commentId, replyingToID: replyingToID), sender: sender)
             }
             actions.append(deleteAction)
         }
@@ -217,7 +217,9 @@ final class SPMainConversationModel {
         let cancelAction = UIAlertAction(
             title: LocalizationManager.localizedString(key: "Cancel"),
             style: .cancel
-        )
+        ) { action in
+            SPAnalyticsHolder.default.log(event: .messageContextMenuClosed(messageId: commentId, relatedMessageId: replyingToID), source: .conversation)
+        }
         if !actions.isEmpty {
             actions.append(cancelAction)
         }
@@ -375,7 +377,7 @@ extension SPMainConversationModel: RealTimeServiceDelegate {
     }
 }
 
-protocol CommentsActionDelegate: class {
+protocol CommentsActionDelegate: AnyObject {
     
     func prepareFlowForAction(_ type: ActionType, sender: UIButton)
     func localCommentWasCreated()
@@ -384,8 +386,8 @@ protocol CommentsActionDelegate: class {
 }
 
 enum ActionType {
-    case delete(commentId: String)
-    case report(commentId: String)
-    case edit(commentId: String)
-    case share(commentId: String)
+    case delete(commentId: String, replyingToID: String?)
+    case report(commentId: String, replyingToID: String?)
+    case edit(commentId: String, replyingToID: String?)
+    case share(commentId: String, replyingToID: String?)
 }